backend/server/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0.1
django-cors-headers==4.3.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
pillow==10.2.0
pycparser==2.21
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
